refactor: Add routers to search and settings
Added routers for search and settings endpoints

Refactor #COG-334
