+++ cognee/api/v1/search/routers/__init__.py
@@ -1,0 +1,1 @@
+from .get_search_router import get_search_router
